+++ abogen/tts_backends/kokoro.py
@@ -1,0 +1,5 @@
+def load_numpy_kpipeline():
+    import numpy as np
+    from kokoro import KPipeline  # type: ignore[import-not-found]
+
+    return np, KPipeline
