+++ tests/test_tts_supertonic_unsupported_chars.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from abogen.tts_supertonic import SupertonicPipeline
+from abogen.tts_backends.supertonic import SupertonicPipeline
 
 
 class _DummyTTS:
